Content preparation
Create a public folder and place the static files inside it.
Write the appropriate middleware function to register the public folder with the application.
Install the EJS module.
Specify in the application that we will use the EJS module.
Create a views folder and convert all .html files within it into .ejs files.
Create the partials folder structure.
Set up the necessary routes and update the links in the _navigation.ejs file.